blocks: rename multiply_const (remove blk)
