+++ schedulers/simplestream/test/test_one_to_many.cpp
@@ -5,7 +5,7 @@
 #include <gnuradio/blocklib/blocks/throttle.hpp>
 #include <gnuradio/blocklib/blocks/fanout.hpp>
 #include <gnuradio/blocklib/blocks/vector_sink.hpp>
-#include <gnuradio/blocklib/blocks/multiply_const_blk.hpp>
+#include <gnuradio/blocklib/blocks/multiply_const.hpp>
 #include <gnuradio/blocklib/blocks/vector_source.hpp>
 #include <gnuradio/flowgraph.hpp>
 #include <gnuradio/schedulers/simplestream/scheduler_simplestream.hpp>
