+++ schedulers/simplestream/test/test_parameters_callbacks.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <thread>
 
-#include <gnuradio/blocklib/blocks/multiply_const_blk.hpp>
+#include <gnuradio/blocklib/blocks/multiply_const.hpp>
 #include <gnuradio/blocklib/blocks/throttle.hpp>
 #include <gnuradio/blocklib/blocks/vector_sink.hpp>
 #include <gnuradio/blocklib/blocks/vector_source.hpp>
